+++ previous-versions/commerce/src/codatcommerce/models/operations/get_tax_component.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import taxcomponent as shared_taxcomponent
+from ...models.shared import taxcomponent as shared_taxcomponent
 from typing import Optional
 
 
@@ -24,12 +23,10 @@ class GetTaxComponentRequest:
 class GetTaxComponentResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     tax_component: Optional[shared_taxcomponent.TaxComponent] = dataclasses.field(default=None)
     r"""OK"""